samples/agent/local_llm/pyproject.toml
[project]
name = "a2ui-local-agent"
version = "0.1.0"
description = "Multi-backend A2UI agent with pluggable LLM providers"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["a2ui", "agent", "llm", "gemini"]
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi[standard]",
    "httpx",
    "orjson",
    "jsonschema>=4.0.0",
    "a2ui-agent",
    "google-genai>=1.27.0",
    "sse-starlette>=2.0.0",
]

[project.scripts]
a2ui-local-agent = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.uv.sources]
a2ui-agent = { path = "../../../agent_sdks/python", editable = true }

[tool.pyright]
include = ["."]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff]
line-length = 120
target-version = "py310"

[dependency-groups]
dev = ["pyright>=1.1.378", "ruff>=0.6.9"]
